fix github comments
