+++ accordion/src/components/AccordionForm/AccordionForm.tsx
@@ -2,13 +2,14 @@ import React, { useCallback, useState } from 'react';
 import AccordionList from '../AccordionList/AccordionList';
 import { AccordionType } from '../../Types';
 import styled from 'styled-components';
+import { searchAccordion } from './AccordionFormFunctions';
 
-const AccordionDiv = styled.div`
+const AccordionWrapper = styled.div`
   border: 1px solid black;
   border-radius: 2%;
   padding: 5px;
   background-color: #d1d1d1;
-  margin: 10px;
+  margin-top: 10px;
   box-shadow:
     0 14px 28px rgba(0, 0, 0, 0.25),
     0 10px 10px rgba(0, 0, 0, 0.22);
@@ -47,20 +48,6 @@ const AccordionForm: React.FC<AccordionFormProps> = ({ data }) => {
     });
   }, []);
 
-  const searchAccordion = useCallback((data: AccordionType[], search: string): AccordionType[] => {
-    return data.map((accordion) => {
-      const childrenAccordion = searchAccordion(accordion.children, search);
-      const isOpen =
-        childrenAccordion.some((child) => child.open) ||
-        childrenAccordion.some((child) => child.title.toLowerCase().includes(search.toLowerCase()));
-      return {
-        ...accordion,
-        open: isOpen,
-        children: childrenAccordion,
-      };
-    });
-  }, []);
-
   const handleSearch = useCallback(
     (e: React.ChangeEvent<HTMLInputElement>) => {
       setAccordionSearch(e.target.value);
@@ -76,9 +63,9 @@ const AccordionForm: React.FC<AccordionFormProps> = ({ data }) => {
   return (
     <>
       <Input type="text" placeholder="Поиск..." value={accordionSearch} onChange={handleSearch} />
-      <AccordionDiv>
+      <AccordionWrapper>
         <AccordionList data={accordionTree} handleSwitcher={handleSwitcher} />
-      </AccordionDiv>
+      </AccordionWrapper>
     </>
   );
 };
