+++ accordion/src/components/AccordionList/AccordionList.tsx
@@ -1,6 +1,11 @@
 import React from 'react';
 import { AccordionType } from '../../Types';
 import AccordionItem from '../AccordionItem/AccordionItem';
+import styled from 'styled-components';
+
+const AccordionChildren = styled.div`
+  padding: 0 0 0 15px;
+`;
 
 interface AccordionListProps {
   data: AccordionType[];
@@ -12,7 +17,9 @@ const AccordionList: React.FC<AccordionListProps> = ({ data, handleSwitcher }) =
     <>
       {data.map((accordion) => (
         <AccordionItem key={accordion.id} accordion={accordion} handleSwitcher={handleSwitcher}>
-          <AccordionList data={accordion.children} handleSwitcher={handleSwitcher} />
+          <AccordionChildren>
+            <AccordionList data={accordion.children} handleSwitcher={handleSwitcher} />
+          </AccordionChildren>
         </AccordionItem>
       ))}
     </>
